requirements.txt
onnxruntime-gpu==1.21.0
dghs-imgutils @ git+https://github.com/deepghs/imgutils@main
waifuc[gpu] @ git+https://github.com/deepghs/waifuc.git@main
python-dotenv==1.1.0
matplotlib==3.10.1
boto3==1.37.23
gradio>=4.0.0